refactor: Map 객체 제거
